requirements.txt
bleach==1.5.0
decorator==4.0.10
entrypoints==0.2.2
html5lib==0.9999999
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
lxml==3.7.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==5.0.0
nbformat==4.2.0
notebook==4.2.0
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
python-docx==0.8.6
pyzmq==16.0.2
qtconsole==4.2.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
wheel==0.29.0
widgetsnbextension==1.2.6
